refactor(navigation): Update navigateTo method to accept speed multiplier and adjust BotMoveTaskParams to use float

+++ src/main/java/com/devone/bot/core/logic/tasks/move/params/BotMoveTaskParams.java
@@ -6,19 +6,19 @@ import com.devone.bot.utils.blocks.BotCoordinate3D;
 public class BotMoveTaskParams extends BotTaskParams { 
     // 🆕 Параметры задачи движения
     public BotCoordinate3D target;
-    public double speedMultiplier;
+    public float speedMultiplier;
 
-    public BotMoveTaskParams(BotCoordinate3D target, double speedMultiplier) {
+    public BotMoveTaskParams(BotCoordinate3D target, float speedMultiplier) {
         this.target = target;
         this.speedMultiplier = speedMultiplier;
     }
     public BotMoveTaskParams(BotCoordinate3D target) {
         this.target = target;
-        this.speedMultiplier = 1.0;
+        this.speedMultiplier = 1.0F;
     }
     public BotMoveTaskParams() {
         this.target = null;
-        this.speedMultiplier = 1.0;
+        this.speedMultiplier = 1.0F;
     }
     public BotCoordinate3D getTarget() {
         return target;
@@ -29,7 +29,7 @@ public class BotMoveTaskParams extends BotTaskParams {
     public double getSpeedMultiplier() {
         return speedMultiplier;
     }
-    public void setSpeedMultiplier(double speedMultiplier) {
+    public void setSpeedMultiplier(float speedMultiplier) {
         this.speedMultiplier = speedMultiplier;
     }
     @Override
